refactor: image files take real android gallery name

+++ app/src/main/java/com/ifyezedev/coslog/core/common/usecase/SaveBitmapsToInternalStorage.kt
@@ -2,7 +2,6 @@ package com.ifyezedev.coslog.core.common.usecase
 
 import android.graphics.Bitmap
 import android.net.Uri
-import androidx.core.net.toUri
 import com.ifyezedev.coslog.core.exception.Failure
 import com.ifyezedev.coslog.core.functional.Either
 import com.ifyezedev.coslog.feature.elements.internal.FilePathProvider
@@ -16,7 +15,6 @@ class SaveBitmapsToInternalStorage(
 
     override suspend fun run(params: List<Pair<Uri, Bitmap>>): Either<List<String>, Failure> {
         val paths = mutableListOf<String>()
-
         // make sure the image directory exists
         val file = File(filePathProvider.getDefaultImageDirectory())
         if (!file.exists())
@@ -24,13 +22,18 @@ class SaveBitmapsToInternalStorage(
 
         params.forEach { bitmapPathPair ->
             try {
-                val path = filePathProvider.generateFilePath(bitmapPathPair.first)
+                val path = filePathProvider.getInternalStorageFilePath(bitmapPathPair.first)
+                val newFile = File(path)
+                paths.add(path)
+
+                if(newFile.exists()) {
+                    return@forEach
+                }
 
                 FileOutputStream(File(path)).use { stream ->
                     bitmapPathPair.second.compress(Bitmap.CompressFormat.PNG, 90, stream)
                 }
 
-                paths.add(path)
             } catch (ex: FileNotFoundException) {
                 return Either.Failure(Failure.IOError(ex.message!!))
             }
